+++ src/Epr.Reprocessor.Exporter.UI/Enums/AddressOptions.cs
@@ -5,12 +5,17 @@ namespace Epr.Reprocessor.Exporter.UI.Enums;
 
 public enum AddressOptions
 {
+    None = 0,
+
     [Description("The registered address")]
     [Display(Name = "The registered address")]
     RegisteredAddress = 1,
+
     [Description("The site address")]
     [Display(Name = "The site address")]
+
     SiteAddress = 2,
+
     [Description("It's a different address")]
     [Display(Name = "It's a different address")]
     DifferentAddress = 3
