+++ src/Epr.Reprocessor.Exporter.UI/Sessions/RegistrationApplicationSession.cs
@@ -1,0 +1,16 @@
+﻿using System.Diagnostics.CodeAnalysis;
+using Epr.Reprocessor.Exporter.UI.Domain;
+
+namespace Epr.Reprocessor.Exporter.UI.Sessions;
+
+/// <summary>
+/// Represents a session for the registration application.
+/// </summary>
+[ExcludeFromCodeCoverage]
+public class RegistrationApplicationSession
+{
+    /// <summary>
+    /// Details of the reprocessing site.
+    /// </summary>
+    public ReprocessingSite? ReprocessingSite { get; set; } = new();
+}
