+++ src/Epr.Reprocessor.Exporter.UI/Views/Registration/AddressOfReprocessingSite.cshtml
@@ -44,7 +44,7 @@
 							   id="SelectedOption"
 							   name="SelectedOption"
 							   type="radio"
-							   value="@AddressOptions.SiteAddress"
+							   value="@AddressOptions.SiteAddress" 
 							   aria-controls="conditional-address" />
 
 						<label class="govuk-label govuk-radios__label" for="SelectedOption">
